fix(execution): live 差分买卖、减仓豁免、撤单容错与拒绝报告边界
三份规则审查的修复批次：
- live 执行器改按 READ_ONLY 持仓差分折算委托（目标低于持仓卖出），此前
  从零折算会重复买入且永不卖出；买入前核可用 JPY 含手续费余量
- 信封门禁：卖出减仓单豁免总额、当日额、单笔上限与首单 canary 门
- 撤单异常不再崩溃；发送超时先等落账再查询；连续写异常随信封状态持久
  化并注入熔断器；用量文件从账本补记；私有客户端共用限速器
- live 拒绝调试覆盖参数；观察心跳超龄拒绝进入；接口错误、网络超时与
  目标校验失败转为拒绝报告；写路径 HTTP 5xx 视为结果未知；时间戳必须带
  时区；观察进程单轮失败不退出
- 数据层：合并失败先回滚；K 线物化取单写锁；启动器设 UTF-8 输出
- 研究链：闭环强制搜索截止与研究截止一致；结构 challenger 计入试验数；
  相关矩阵双精度；DSR 离散度取样本外

触碰端点：无新增写端点；撤单失败路径仍只经 POST /v1/cancelOrder（A-03）。

+++ src/guvolu/search/resample.py
@@ -569,7 +569,8 @@ def standardize_rows(torch: object, values: Tensor) -> tuple[Tensor, int]:
     norm = centered.pow(2).sum(dim=1).sqrt()
     constant = norm <= 0.0
     safe = norm.where(~constant, norm.new_ones(()))
-    standardized = (centered / safe.unsqueeze(1)).to(values.new_zeros(()).float().dtype)
+    # 双精度：相关矩阵特征值对精度敏感
+    standardized = (centered / safe.unsqueeze(1)).to(values.new_zeros(()).double().dtype)
     standardized[constant] = 0.0
     return standardized, int(constant.sum().item())
